editor: refactor clipboard extension
