+++ packages/editor/package.json
@@ -33,7 +33,6 @@
     "@tiptap/extension-underline": "2.0.3",
     "@tiptap/pm": "2.0.3",
     "@tiptap/starter-kit": "2.0.3",
-    "clipboard-polyfill": "4.0.0",
     "detect-indent": "^7.0.0",
     "entities": "^4.5.0",
     "katex": "0.16.4",
