+++ packages/editor/src/extensions/clipboard/clipboard.ts
@@ -16,8 +16,15 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
+
 import { Extension } from "@tiptap/core";
-import { writeText } from "clipboard-polyfill";
+import { Plugin, PluginKey } from "prosemirror-state";
+import { Slice } from "prosemirror-model";
+import { LIST_NODE_TYPES } from "../../utils/node-types";
+import { ClipboardDOMParser } from "./clipboard-dom-parser";
+import { ClipboardDOMSerializer } from "./clipboard-dom-serializer";
+import { clipboardTextParser } from "./clipboard-text-parser";
+import { clipboardTextSerializer } from "./clipboard-text-serializer";
 
 declare module "@tiptap/core" {
   interface Commands<ReturnType> {
@@ -31,14 +38,15 @@ export type ClipboardOptions = {
   copyToClipboard: (text: string) => void;
 };
 
-export const Clipboard = Extension.create<ClipboardOptions>({
+export const Clipboard = Extension.create({
+  name: "clipboard",
+
   addOptions() {
     return {
-      copyToClipboard: (text) => {
-        writeText(text);
-      }
+      copyToClipboard: () => {}
     };
   },
+
   addCommands() {
     return {
       copyToClipboard: (text: string) => (props) => {
@@ -46,5 +54,39 @@ export const Clipboard = Extension.create<ClipboardOptions>({
         return true;
       }
     };
+  },
+
+  addProseMirrorPlugins() {
+    return [
+      new Plugin({
+        key: new PluginKey("clipboard"),
+        props: {
+          clipboardParser: ClipboardDOMParser.fromSchema(
+            this.editor.view.state.schema
+          ),
+          clipboardSerializer: ClipboardDOMSerializer.fromSchema(
+            this.editor.view.state.schema
+          ),
+          transformCopied,
+          clipboardTextParser,
+          clipboardTextSerializer
+        }
+      })
+    ];
   }
 });
+
+export function transformCopied(slice: Slice) {
+  // when copying a single list item, we shouldn't retain the
+  // list formatting but copy it as a paragraph.
+  const maybeList = slice.content.firstChild;
+  if (
+    maybeList &&
+    LIST_NODE_TYPES.includes(maybeList.type.name) &&
+    maybeList.childCount === 1 &&
+    maybeList.firstChild
+  ) {
+    return transformCopied(new Slice(maybeList.firstChild.content, 0, 0));
+  }
+  return slice;
+}
