+++ packages/editor/src/extensions/clipboard/index.ts
@@ -16,8 +16,6 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
-import { Clipboard } from "./clipboard";
 
 export * from "./clipboard";
-
-export default Clipboard;
+export { Clipboard as default } from "./clipboard";
